Show Multiple Phone Numbers of the same type
If a person has multiple of the same phone type, all numbers of the type will be displayed on the edit page, and the user would be able to successfully edit or remove one of numbers successfully.  Previous behavior was that only one phone number of each type was displayed.

(cherry picked from commit 7b2497bf90b0730af277d62d641f65349a33acf0)